+++ test/pytests/test_smart_completion_public_schema_only.py
@@ -99,7 +99,10 @@ def test_special_name_completion(completer, complete_event):
     text = "\\d"
     position = len("\\d")
     result = completer.get_completions(Document(text=text, cursor_position=position), complete_event)
-    assert list(result) == [Completion(text="\\dt", start_position=-2)]
+    assert list(result) == [
+        Completion(text="\\dt", start_position=-2),
+        Completion(text="\\dsn", start_position=-2),
+    ]
 
 
 def test_empty_string_completion(completer, complete_event):